requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2.3
django-allauth==0.54.0
django-cache-url==3.4.4
django-environ==0.10.0
environs==9.5.0
gunicorn==21.2.0
idna==3.4
marshmallow==3.20.1
oauthlib==3.2.2
packaging==23.1
paramiko==3.3.1
psycopg==3.1.10
psycopg-binary==3.1.10
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.5.0
